+++ msmbuilder/tests/test_feature_selection.py
@@ -1,8 +1,9 @@
 import numpy as np
+import pandas as pd
 from sklearn.feature_selection import VarianceThreshold as VarianceThresholdR
 
 from msmbuilder.example_datasets import AlanineDipeptide
-from msmbuilder.feature_selection import FeatureSelector, VarianceThreshold
+from msmbuilder.feature_selection import FeatureSelector, VarianceThreshold, FeatureSlicer
 from msmbuilder.featurizer import DihedralFeaturizer
 
 FEATS = [
@@ -73,3 +74,24 @@ def test_which_feat_types():
         pass
 
 
+
+def test_feature_slicer():
+    trajectories = AlanineDipeptide().get_cached().trajectories
+    f = DihedralFeaturizer()
+    fs = FeatureSlicer(f, indices=[0,1])
+    y1 = fs.transform(trajectories)
+    assert y1[0].shape[1] == 2
+
+    df = pd.DataFrame(fs.describe_features(trajectories[0]))
+    assert len(df) == 2
+    assert 'psi' not in df.featuregroup[0]
+    assert 'psi' not in df.featuregroup[1]
+
+    fs = FeatureSlicer(f, indices=[2,3])
+    y1 = fs.transform(trajectories)
+    assert y1[0].shape[1] == 2
+
+    df = pd.DataFrame(fs.describe_features(trajectories[0]))
+    assert len(df) == 2
+    assert 'phi' not in df.featuregroup[0]
+    assert 'phi' not in df.featuregroup[1]