+++ docs/changelog.rst
@@ -11,9 +11,12 @@ API Changes
 
 New Features
 ~~~~~~~~~~~~
+- Added new featurizer ```FeatureSlicer```. ```FeatureSlicer``` can slice the ouput of
+regular featurizer objects to just the required indices.
 - Added functions to compute error bars for transition probabilities to account for
   finite sampling, and sample transition matrices from these error distributions (i.e.
   bootstrapping). Located in ```msmbuilder.msm.validation.transmat_errorbar```.
+
 
 Improvements
 ~~~~~~~~~~~~